Luuna/requirements.txt
asgiref==3.2.10
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-filter==21.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.0.4
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonfield==3.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
sqlparse==0.4.1
swapper==1.1.2.post1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
whitenoise==5.2.0
